pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nightingale"
version = "0.1.0"
description = "Autonomous SRE Agent"
authors = [
  { name="User", email="user@example.com" },
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0",
    "PyYAML>=6.0",
    "gitpython>=3.1.0",
    "rich>=13.0",
    "pytest>=7.0",
    "google-generativeai>=0.3.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["nightingale*"]
